requirements.txt
streamlit
faiss-cpu
sentence-transformers
pdfplumber
ollama
python-dotenv
fpdf